backend/pyproject.toml
[project]
name = "langly-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "asyncpg>=0.30.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "PyJWT>=2.8.0",
    "cryptography>=42.0.0",
    "httpx>=0.27.0",
]
